# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/boards/{board}/columns": {
            "get": {
                "tags": [
                    "Columns"
                ],
                "summary": "Obtener todas las columnas de un tablero",
                "description": "Retorna la lista de todas las columnas de un tablero específico, incluyendo sus tareas ordenadas",
                "operationId": "aceff0968a12ce5d535eeb9629062835",
                "parameters": [
                    {
                        "name": "board",
                        "in": "path",
                        "description": "ID del tablero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Columnas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "board_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Pendiente"
                                            },
                                            "order": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "color": {
                                                "type": "string",
                                                "example": "#ba181b"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "tasks": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Columns"
                ],
                "summary": "Crear una nueva columna en un tablero",
                "description": "Crea una nueva columna en el tablero especificado. Si no se proporciona el orden, se asigna automáticamente al final.",
                "operationId": "b181e04fde144c7f5c3de11ec657d9fe",
                "parameters": [
                    {
                        "name": "board",
                        "in": "path",
                        "description": "ID del tablero donde se creará la columna",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre de la columna",
                                        "type": "string",
                                        "example": "En Revisión"
                                    },
                                    "color": {
                                        "description": "Color de la columna en formato HEX",
                                        "type": "string",
                                        "example": "#ffa500"
                                    },
                                    "order": {
                                        "description": "Posición de la columna en el tablero",
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Columna creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Column created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "board_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "En Revisión"
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "#ffa500"
                                                },
                                                "order": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create column"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}